Deprecate CollectionRequest and remove unused constructors
CollectionRequest tests that have not been superceded by the mixin
test suite have been moved to WPRequest
